requirements.txt
streamlit==1.23.1
firebase-admin==6.0.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
python-dotenv
streamlit-option-menu
PyPDF2
langchain
langchain_community
langchain-google-genai
pymongo


# gcloud init --skip-diagnostics